// reality/cloud/xrhome/src/client/common/types/db.ts
type AppCommercialStatus = "DEMO" | "DEVELOP" | "LAUNCH" | "COMPLETE" | "CANCELED"
type AppRepoStatus = "UNKNOWN" | "NONE" | "PRIVATE" | "PUBLIC"
type AppBuildSettingsSplashScreen = "UNKNOWN" | "DEFAULT" | "DISABLED" | "DEMO" | "NONCOMMERCIAL" | "EDUCATIONAL" | "APP"
// ...
type UserAccountRole = "OWNER" | "ADMIN" | "DEV" | "BILLMANAGER"